feat(test-tools): rename module
