+++ libs/test-tools/jest-e2e.json
@@ -11,7 +11,7 @@
     "^@apps/gateway/(.*)$": "<rootDir>/apps/gateway/src/$1",
     "^@libs/health-check": "<rootDir>/libs/health-check/src/index.ts",
     "^@libs/health-check/(.*)$": "<rootDir>/libs/health-check/src/$1",
-    "^@libs/test-helper": "<rootDir>/libs/test-helper/src/index.ts",
-    "^@libs/test-helper/(.*)$": "<rootDir>/libs/test-helper/src/$1"
+    "^@libs/test-tools": "<rootDir>/libs/test-tools/src/index.ts",
+    "^@libs/test-tools/(.*)$": "<rootDir>/libs/test-tools/src/$1"
   }
 }
