+++ package.json
@@ -22,7 +22,7 @@
     "test:watch": "jest --watch",
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
-    "test:e2e": "jest --config ./libs/test-helper/jest-e2e.json"
+    "test:e2e": "jest --config ./libs/test-tools/jest-e2e.json"
   },
   "dependencies": {
     "@nestjs/common": "^9.0.0",
