+++ tsconfig.json
@@ -20,8 +20,8 @@
     "paths": {
       "@apps/sw/*": ["apps/sw/src/*"],
       "@apps/gateway/*": ["apps/gateway/src/*"],
-      "@libs/test-helper": ["libs/test-helper/src"],
-      "@libs/test-helper/*": ["libs/test-helper/src/*"],
+      "@libs/test-tools": ["libs/test-tools/src"],
+      "@libs/test-tools/*": ["libs/test-tools/src/*"],
       "@libs/health-check": ["libs/health-check/src"],
       "@libs/health-check/*": ["libs/health-check/src/*"],
       "@libs/abstract-tools": ["libs/abstract-tools/src"],
